CRUD Wallet FIX with nice admin

+++ database/migrations/2025_01_15_133313_create_wallets_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('wallets', function (Blueprint $table) {
             $table->id();
             $table->string('name');
-            $table->decimal('jumlah', 15, 2);
+            $table->integer('jumlah');
             $table->timestamps();
         });
     }
